Fix 'psecrets unlock', simplify argument parsing
Fix the secrets repo so that we can actually push commits we make.
Previously, we attempted to have an encrypted secrets submodule, then
clone from that submodule inside the container into a decrypted
checkout directory. The plan was to push from the decrypted checkout
directory back to the encrypted secrets submodule inside the container,
and then push from the encrypted secrets submodule to the real git
remote outside the container. Unfortunately, this doesn't work; the
submodule doesn't detect the changes. Instead, just clone the secrets
repo when the container starts. Slower, but at least it actually works.

Also, simplify argument parsing so that all arguments are available for
all subparsers.

- Mount $PSYOPS_SECRETS_PATH without default 'noexec'
- Change secrets repo URI
- Always clone secrets repo on login - unfortunate because its slow
- Fix ln call for private SSH key
- Modify secretssetup() function