Hack around the lack of /dev/disk/by-path
Ultimately, this is a bad idea, because it can't work in the Ceph
configuration, so I'm recording this just to scrap it.

Change WholeDiskSpec and PartitionSpec to accept /sys paths to device
information directories. Convert these to /dev paths.
This only works on the target host with the live /sys filesystem of
course, but worse, it cannot be used in Ceph configuration files.
The Ceph config in the cluster needs a /dev device path, and it can't
accept /sys paths like we do.

What I really need is /dev/disk/by-path. I'm not sure if I can get that
with mdev, or if I'll have to switch to udev.