Fix version/apk problems during build
* Fix build errors by setting tool.setuptools.packages in pyproject.toml
* Set the APK repository name by creating a tmpdir
* Add lots of debugging

Previously, we had set tool.setuptools.dynamic.version.attr in
pyproject.toml, but it wasn't working properly.

In particular, progfigsite.get_version() would get a ModuleNotFoundError
when it tried to import progfigsite.builddata.version.
After a long debugging process,
which resulted in some new tooling,
I discovered that setuptools needs tool.setuptools.packages set for this
to work properly in my case.

This was obfuscated by the fact that sometimes I had installed
progfigsite into PYTHONPATH with 'pip install -e .'.
It was mostly working, but I failed to realize that it was looking for
progfigsite.builddata.version **in PYTHONPATH**, not in its package dir.
Setting tool.setuptools.packages fixes this.

In the course of this work I also realized that abuild uses the parent
directory name for the APK repository name and there's no other way to
set it. To work around this limitation, we now copy everything to a temp
dir with the name we want to use for the APK repo before building the
APK.