Improve grubusb squashfs
- Completely remove old grubusbsq stuff
- Create grubusb squashfs in a separate script/build step
- Create a sparse grubusb image to save space and time

The grubusbsq method was generating a squashfs filesystem with
mkimage.sh. This reused some code, but we need to customize our
initramfs-init no matter what to find the right A/B partition at boot,
so this was only workable for a demo. (Also, this means grubusb images
can be built without an aports checkout.)